requirements.txt
requests>=2.32
selenium>=4.0
curl_cffi>=0.7
pycryptodome>=3.20
